requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
openai>=1.65.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
chromadb>=0.6.3
sentence-transformers>=3.4.0
python-dotenv>=1.0.1
Pillow>=11.1.0
